Refactor DTO schema helpers
